practice_project/pyproject.toml
[project]
name = "ai-spec-driven-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "uvicorn>=0.38.0",
]
